prediksi-cuaca/requirements.txt
streamlit>=1.28.0
folium==0.15.1
streamlit-folium==0.17.0
requests>=2.31.0
pandas>=2.0.0
altair>=5.0.0
plotly>=5.17.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0
pmdarima>=2.0.0
prophet>=1.1.0
tensorflow-cpu>=2.15.0
xgboost>=2.0.0
scikit-learn>=1.3.0
